Handle links-for-invalid-sites comment submit
It is possible for links to match the `links` regex but whose domain does not match the `validSites` regex.

In such cases, the comment box is never submitted, and a second attempt to do so results in a page reload. This change, checking for the lock status in the handler lets you detect both the 'no links found' and the 'no links for valid sites' cases, and submit the comment form directly.